Add a relation users to channels saga
