+++ src/store/channels-list/index.ts
@@ -2,19 +2,16 @@ import { createAction } from '@reduxjs/toolkit';
 import { createNormalizedListSlice } from '../normalized';
 
 import { schema } from '../channels';
-import { Payload } from './saga';
 
 export enum SagaActionTypes {
   Fetch = 'channelsList/saga/fetch',
   ReceiveUnreadCount = 'channelsList/saga/receiveUnreadCount',
   StopSyncChannels = 'channelsList/saga/stopSyncChannels',
-  LoadUsers = 'channelsList/saga/loadUsers',
 }
 
 const fetch = createAction<string>(SagaActionTypes.Fetch);
 const receiveUnreadCount = createAction<string>(SagaActionTypes.ReceiveUnreadCount);
 const stopSyncChannels = createAction<string>(SagaActionTypes.StopSyncChannels);
-const loadUsers = createAction<Payload>(SagaActionTypes.LoadUsers);
 
 const slice = createNormalizedListSlice({
   name: 'channelsList',
@@ -23,4 +20,4 @@ const slice = createNormalizedListSlice({
 
 export const { receiveNormalized, setStatus, receive } = slice.actions;
 export const { reducer, normalize, denormalize } = slice;
-export { fetch, receiveUnreadCount, stopSyncChannels, loadUsers };
+export { fetch, receiveUnreadCount, stopSyncChannels };
