+++ src/store/channels-list/saga.ts
@@ -2,7 +2,7 @@ import getDeepProperty from 'lodash.get';
 import { takeLatest, put, call, delay } from 'redux-saga/effects';
 import { SagaActionTypes, setStatus, receive } from '.';
 
-import { fetchChannels, fetchUsersByChannelId } from './api';
+import { fetchChannels } from './api';
 import { AsyncListStatus } from '../normalized';
 import { select } from 'redux-saga-test-plan/matchers';
 
@@ -47,24 +47,6 @@ export function* unreadCountUpdated(action) {
   yield put(receive(channelsList));
 }
 
-export function* loadUsers(action) {
-  const { channelId } = action.payload;
-  const channelPrefix: string = channelIdPrefix + channelId;
-
-  const users = yield call(fetchUsersByChannelId, channelPrefix);
-  const formatUsers = users.map(({ userId: id, ...rest }) => ({
-    id,
-    ...rest,
-  }));
-
-  yield put(
-    receive({
-      id: channelId,
-      users: formatUsers,
-    })
-  );
-}
-
 export function* stopSyncChannels() {
   yield put(setStatus(AsyncListStatus.Stopped));
 }
@@ -80,5 +62,4 @@ export function* saga() {
   yield takeLatest(SagaActionTypes.Fetch, fetch);
   yield takeLatest(SagaActionTypes.ReceiveUnreadCount, syncUnreadCount);
   yield takeLatest(SagaActionTypes.StopSyncChannels, stopSyncChannels);
-  yield takeLatest(SagaActionTypes.LoadUsers, loadUsers);
 }
