+++ src/store/channels/api.ts
@@ -1,0 +1,7 @@
+import { get } from '../../lib/api/rest';
+
+export async function fetchUsersByChannelId(channelId: string): Promise<any> {
+  const response = await get<any>(`/chatChannels/${channelId}/members`);
+
+  return response.body;
+}
